ocfemia-lizelle-2023-hw5/requirements.txt
matplotlib
pyyaml
tqdm
scikit-learn
datasets
sentence_transformers